refactor : domain 디렉토리 분리
