+++ src/main/java/com/example/hospitalreservation/domain/Doctor.java
@@ -1,4 +1,4 @@
-package com.example.hospitalreservation.model;
+package com.example.hospitalreservation.domain;
 
 public class Doctor {
     private Long id;
