+++ src/main/java/com/example/hospitalreservation/repository/ReservationRepository.java
@@ -1,6 +1,6 @@
 package com.example.hospitalreservation.repository;
 
-import com.example.hospitalreservation.model.Reservation;
+import com.example.hospitalreservation.domain.Reservation;
 import org.springframework.stereotype.Repository;
 
 import java.util.ArrayList;
